Introduce `MigrateImageMessage` (and moved `InternalThrallMessage` and `ExternalThrallMessage` into `ThrallMessage` to ensure non-exhaustive pattern match warnings occur)
