--- common-lib/src/main/scala/com/gu/mediaservice/model/InternalThrallMessage.scala
@@ -1,6 +1,0 @@
-package com.gu.mediaservice.model
-
-sealed trait InternalThrallMessage extends ThrallMessage {
-
-}
-
